Refactor: refactor dei repository tramite interfaccia per separare maggiormente le funzionalità.
